// src/types/json-schema.ts
export enum SchemaRendererVariant {
  Sections = 'sections',
  Flat = 'flat',
  FlatSections = 'flat-sections',
}

export enum SchemaDisplayMode {
  Select = 'select',
  Radio = 'radio',
}

export enum SchemaOrientation {
  Row = 'row',
  Column = 'column',
}

export enum JsonSchemaType {
  String = 'string',
  Number = 'number',
  Integer = 'integer',
  Boolean = 'boolean',
  Object = 'object',
  Array = 'array',
  Null = 'null',
}